..F....... [ZBX-18975] fixed item value being truncated in item test form
* commit 'b535eb8dc59781742d9181a3456e73c7a58d70a1':
  ..F....... [ZBX-18975] recompiled css
  ..F....... [ZBX-18975] made maxlength attribute optional
  ..F....... [ZBX-18975] made maxlength attribute optional
  .D........ [ZBX-18975] added changelog file
  ..F....... [ZBX-18975] fixed item value being truncated in item test form

(cherry picked from commit 898ea5ed959425dda3a885b86de0f22a4fad1149)

+++ ui/include/views/js/item.preprocessing.js.php
@@ -185,7 +185,7 @@
 						label_after: '}',
 						grow: 'auto',
 						rows: 0,
-						maxlength: <?= (int) DB::getFieldLength('item_preproc', 'params') ?>
+						maxlength: <?= DB::getFieldLength('item_preproc', 'params') ?>
 					});
 
 				case '<?= ZBX_PREPROC_PROMETHEUS_PATTERN ?>':
